project_21_website_qa/requirements.txt
langchain
langchain-google-genai
python-dotenv
beautifulsoup4
lxml
requests